Make HWLOC global for all experiments
