+++ lib/benchpark/experiment.py
@@ -5,6 +5,7 @@
 
 from typing import Dict
 import yaml  # TODO: some way to ensure yaml available
+from enum import Enum
 
 from benchpark.error import BenchparkError
 from benchpark.directives import ExperimentSystemBase
@@ -145,7 +146,34 @@ class Affinity:
             }
 
 
-class Experiment(ExperimentSystemBase, SingleNode, Affinity):
+class HwlocVariantValues(str, Enum):
+    NONE = "none"
+    ON = "on"
+
+
+class Hwloc:
+    variant(
+        "hwloc",
+        default=HwlocVariantValues.NONE.value,
+        values=tuple(v.value for v in HwlocVariantValues),
+        multi=False,
+        description="Get underlying infrastructure topology",
+    )
+
+    class Helper(ExperimentHelper):
+        def compute_modifiers_section(self):
+            modifier_list = []
+
+            if not self.spec.satisfies(f"hwloc={HwlocVariantValues.NONE.value}"):
+                affinity_modifier_modes = {}
+                affinity_modifier_modes["name"] = "hwloc"
+                affinity_modifier_modes["mode"] = self.spec.variants["hwloc"][0]
+                modifier_list.append(affinity_modifier_modes)
+
+            return modifier_list
+
+
+class Experiment(ExperimentSystemBase, SingleNode, Affinity, Hwloc):
     """This is the superclass for all benchpark experiments.
 
     ***The Experiment class***
